+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import Header from './Header';
-import './App.css';
+import '../css/App.css';
 
 export default class App extends Component {
   render() {
